requirements.txt
fastapi==0.109.2
httpx==0.27.2
httpx-sse==0.4.0
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.60
langchain-groq==0.3.2
langchain-text-splitters==0.3.6
pydantic==2.6.1
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
starlette==0.36.3
uvicorn==0.27.1
pytest==8.0.0
requests==2.31.0
streamlit==1.31.1